requirements.txt
lancedb>=0.7.0
sentence-transformers>=2.2.2
FlagEmbedding>=1.2.10
rank-bm25>=0.2.2
tiktoken>=0.7.0
pymupdf>=1.23.0
requests>=2.32.0
llama-index-core>=0.10.0
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
